requirements.txt
python-dotenv==1.0.0
openai==0.28.0
langchain==0.0.281
supabase==2.0.3
streamlit-extras==0.3.2